requirements.txt
unsloth
transformers>=5.2
trl==0.22.2
datasets
Pillow
accelerate
scikit-learn
huggingface_hub
wandb
requests
